Adding build script.
Interestingly, increasing the table size in order to make the audio clip last longer, the drum beat pulses. I guess that's to be expected.
The pulses are likely the result of the maxima of the sine waves adding ontop of each other. And the pauses in sound are likely a result of maxima and minima cancelling.